Refactored USS rename functionality
